fix(lookup): properly throw for InvalidData
fixes #122, fixes #124

+++ src/IQFeed.CSharpApiClient/Lookup/Historical/Messages/IIntervalMessage.cs
@@ -9,7 +9,7 @@ namespace IQFeed.CSharpApiClient.Lookup.Historical.Messages
         double Low { get; }
         int NumberOfTrades { get; }
         double Open { get; }
-        int PeriodVolume { get; }
+        long PeriodVolume { get; }
         long TotalVolume { get; }
     }
 }
